+++ PowerWater/PowerLib/URP_GI.hlsl
@@ -4,37 +4,37 @@
 
 
 //-----------------------------------------------------------EntityLight
-#define LIGHTMAP_RGBM_MAX_GAMMA     half(5.0)       // NB: Must match value in RGBMRanges.h
-#define LIGHTMAP_RGBM_MAX_LINEAR    half(34.493242) // LIGHTMAP_RGBM_MAX_GAMMA ^ 2.2
+#define LIGHTMAP_RGBM_MAX_GAMMA     float(5.0)       // NB: Must match value in RGBMRanges.h
+#define LIGHTMAP_RGBM_MAX_LINEAR    float(34.493242) // LIGHTMAP_RGBM_MAX_GAMMA ^ 2.2
 
 #ifdef UNITY_LIGHTMAP_RGBM_ENCODING
     #ifdef UNITY_COLORSPACE_GAMMA
         #define LIGHTMAP_HDR_MULTIPLIER LIGHTMAP_RGBM_MAX_GAMMA
-        #define LIGHTMAP_HDR_EXPONENT   half(1.0)   // Not used in gamma color space
+        #define LIGHTMAP_HDR_EXPONENT   float(1.0)   // Not used in gamma color space
     #else
         #define LIGHTMAP_HDR_MULTIPLIER LIGHTMAP_RGBM_MAX_LINEAR
-        #define LIGHTMAP_HDR_EXPONENT   half(2.2)
+        #define LIGHTMAP_HDR_EXPONENT   float(2.2)
     #endif
 #elif defined(UNITY_LIGHTMAP_DLDR_ENCODING)
     #ifdef UNITY_COLORSPACE_GAMMA
-        #define LIGHTMAP_HDR_MULTIPLIER half(2.0)
+        #define LIGHTMAP_HDR_MULTIPLIER float(2.0)
     #else
-        #define LIGHTMAP_HDR_MULTIPLIER half(4.59) // 2.0 ^ 2.2
+        #define LIGHTMAP_HDR_MULTIPLIER float(4.59) // 2.0 ^ 2.2
     #endif
-    #define LIGHTMAP_HDR_EXPONENT half(0.0)
+    #define LIGHTMAP_HDR_EXPONENT float(0.0)
 #else // (UNITY_LIGHTMAP_FULL_HDR)
-    #define LIGHTMAP_HDR_MULTIPLIER half(1.0)
-    #define LIGHTMAP_HDR_EXPONENT half(1.0)
+    #define LIGHTMAP_HDR_MULTIPLIER float(1.0)
+    #define LIGHTMAP_HDR_EXPONENT float(1.0)
 #endif
 
- half3 UnpackLightmapRGBM(half4 rgbmInput, half4 decodeInstructions)
+ float3 UnpackLightmapRGBM(float4 rgbmInput, float4 decodeInstructions)
 {
 #ifdef UNITY_COLORSPACE_GAMMA
     return rgbmInput.rgb * (rgbmInput.a * decodeInstructions.x);
 #else
     // return rgbmInput.rgb * (pow(rgbmInput.a, decodeInstructions.y) * decodeInstructions.x);
     // optimise 
-    half scale = rgbmInput.w;
+    float scale = rgbmInput.w;
     #if defined(UNITY_LIGHTMAP_RGBM_ENCODING)
         scale = scale * scale;
     #endif
@@ -43,13 +43,13 @@
 #endif
 }
 
-half3 UnpackLightmapDoubleLDR(half4 encodedColor, half4 decodeInstructions)
+float3 UnpackLightmapDoubleLDR(float4 encodedColor, float4 decodeInstructions)
 {
     return encodedColor.rgb * decodeInstructions.x;
 }
 
 #ifndef BUILTIN_TARGET_API
-half3 DecodeLightmap(half4 encodedIlluminance, half4 decodeInstructions)
+float3 DecodeLightmap(float4 encodedIlluminance, float4 decodeInstructions)
 {
 #if defined(UNITY_LIGHTMAP_RGBM_ENCODING)
     return UnpackLightmapRGBM(encodedIlluminance, decodeInstructions);
@@ -61,47 +61,47 @@ half3 DecodeLightmap(half4 encodedIlluminance, half4 decodeInstructions)
 }
 #endif
 
-half3 SampleLightmap(half2 uv){
+float3 SampleLightmap(float2 uv){
     #ifdef UNITY_LIGHTMAP_FULL_HDR
     bool encodedLightmap = false;
 #else
     bool encodedLightmap = true;
 #endif
 
-    half4 decodeInstructions = half4(LIGHTMAP_HDR_MULTIPLIER, LIGHTMAP_HDR_EXPONENT, 0.0h, 0.0h);
-    half4 illum = SAMPLE_TEXTURE2D(unity_Lightmap,samplerunity_Lightmap,uv);
+    float4 decodeInstructions = float4(LIGHTMAP_HDR_MULTIPLIER, LIGHTMAP_HDR_EXPONENT, 0.0h, 0.0h);
+    float4 illum = SAMPLE_TEXTURE2D(unity_Lightmap,samplerunity_Lightmap,uv);
     return DecodeLightmap(illum,decodeInstructions);
 }
 
-half4 SampleShadowMask(half2 uv){
+float4 SampleShadowMask(float2 uv){
     return SAMPLE_TEXTURE2D(unity_ShadowMask,samplerunity_ShadowMask,uv);
 }
 
 
 
 //--------------------- IBL
-half3 DecodeHDREnvironment(half4 encodedIrradiance, half4 decodeInstructions)
+float3 DecodeHDREnvironment(float4 encodedIrradiance, float4 decodeInstructions)
 {
     // Take into account texture alpha if decodeInstructions.w is true(the alpha value affects the RGB channels)
-    half alpha = max(decodeInstructions.w * (encodedIrradiance.a - 1.0) + 1.0, 0.0);
+    float alpha = max(decodeInstructions.w * (encodedIrradiance.a - 1.0) + 1.0, 0.0);
 
     // If Linear mode is not supported we can skip exponent part
     return (decodeInstructions.x * PositivePow(alpha, decodeInstructions.y)) * encodedIrradiance.rgb;
 }
 
-half3 BoxProjectedCubemapDirection(half3 reflectionWS, float3 positionWS, float4 cubemapPositionWS, float4 boxMin, float4 boxMax)
+float3 BoxProjectedCubemapDirection(float3 reflectionWS, float3 positionWS, float4 cubemapPositionWS, float4 boxMin, float4 boxMax)
 {
     // Is this probe using box projection?
     if (cubemapPositionWS.w > 0.0f)
     {
         float3 boxMinMax = (reflectionWS > 0.0f) ? boxMax.xyz : boxMin.xyz;
-        half3 rbMinMax = half3(boxMinMax - positionWS) / reflectionWS;
+        float3 rbMinMax = float3(boxMinMax - positionWS) / reflectionWS;
 
-        half fa = half(min(min(rbMinMax.x, rbMinMax.y), rbMinMax.z));
+        float fa = float(min(min(rbMinMax.x, rbMinMax.y), rbMinMax.z));
 
-        half3 worldPos = half3(positionWS - cubemapPositionWS.xyz);
+        float3 worldPos = float3(positionWS - cubemapPositionWS.xyz);
 
-        half3 result = worldPos + reflectionWS * fa;
+        float3 result = worldPos + reflectionWS * fa;
         return result;
     }
     else
@@ -117,18 +117,18 @@ float CalculateProbeWeight(float3 positionWS, float4 probeBoxMin, float4 probeBo
     return saturate(min(weightDir.x, min(weightDir.y, weightDir.z)));
 }
 
-half CalculateProbeVolumeSqrMagnitude(float4 probeBoxMin, float4 probeBoxMax)
+float CalculateProbeVolumeSqrMagnitude(float4 probeBoxMin, float4 probeBoxMax)
 {
-    half3 maxToMin = half3(probeBoxMax.xyz - probeBoxMin.xyz);
+    float3 maxToMin = float3(probeBoxMax.xyz - probeBoxMin.xyz);
     return dot(maxToMin, maxToMin);
 }
 
-half3 CalculateIrradianceFromReflectionProbes(half3 reflectVector, float3 positionWS, half perceptualRoughness)
+float3 CalculateIrradianceFromReflectionProbes(float3 reflectVector, float3 positionWS, float perceptualRoughness)
 {
-    half probe0Volume = CalculateProbeVolumeSqrMagnitude(unity_SpecCube0_BoxMin, unity_SpecCube0_BoxMax);
-    half probe1Volume = CalculateProbeVolumeSqrMagnitude(unity_SpecCube1_BoxMin, unity_SpecCube1_BoxMax);
+    float probe0Volume = CalculateProbeVolumeSqrMagnitude(unity_SpecCube0_BoxMin, unity_SpecCube0_BoxMax);
+    float probe1Volume = CalculateProbeVolumeSqrMagnitude(unity_SpecCube1_BoxMin, unity_SpecCube1_BoxMax);
 
-    half volumeDiff = probe0Volume - probe1Volume;
+    float volumeDiff = probe0Volume - probe1Volume;
     float importanceSign = unity_SpecCube1_BoxMin.w;
 
     // A probe is dominant if its importance is higher
@@ -150,9 +150,9 @@ half3 CalculateIrradianceFromReflectionProbes(half3 reflectVector, float3 positi
     weightProbe0 /= max(totalWeight, 1.0f);
     weightProbe1 /= max(totalWeight, 1.0f);
 
-    half3 irradiance = half3(0.0h, 0.0h, 0.0h);
-    half3 originalReflectVector = reflectVector;
-    half mip = PerceptualRoughnessToMipmapLevel(perceptualRoughness);
+    float3 irradiance = float3(0.0h, 0.0h, 0.0h);
+    float3 originalReflectVector = reflectVector;
+    float mip = PerceptualRoughnessToMipmapLevel(perceptualRoughness);
 
     // Sample the first reflection probe
     if (weightProbe0 > 0.01f)
@@ -161,7 +161,7 @@ half3 CalculateIrradianceFromReflectionProbes(half3 reflectVector, float3 positi
         reflectVector = BoxProjectedCubemapDirection(originalReflectVector, positionWS, unity_SpecCube0_ProbePosition, unity_SpecCube0_BoxMin, unity_SpecCube0_BoxMax);
 #endif // _REFLECTION_PROBE_BOX_PROJECTION
 
-        half4 encodedIrradiance = half4(SAMPLE_TEXTURECUBE_LOD(unity_SpecCube0, samplerunity_SpecCube0, reflectVector, mip));
+        float4 encodedIrradiance = float4(SAMPLE_TEXTURECUBE_LOD(unity_SpecCube0, samplerunity_SpecCube0, reflectVector, mip));
 
 #if defined(UNITY_USE_NATIVE_HDR)
         irradiance += weightProbe0 * encodedIrradiance.rbg;
@@ -176,7 +176,7 @@ half3 CalculateIrradianceFromReflectionProbes(half3 reflectVector, float3 positi
 #ifdef _REFLECTION_PROBE_BOX_PROJECTION
         reflectVector = BoxProjectedCubemapDirection(originalReflectVector, positionWS, unity_SpecCube1_ProbePosition, unity_SpecCube1_BoxMin, unity_SpecCube1_BoxMax);
 #endif // _REFLECTION_PROBE_BOX_PROJECTION
-        half4 encodedIrradiance = half4(SAMPLE_TEXTURECUBE_LOD(unity_SpecCube1, samplerunity_SpecCube1, reflectVector, mip));
+        float4 encodedIrradiance = float4(SAMPLE_TEXTURECUBE_LOD(unity_SpecCube1, samplerunity_SpecCube1, reflectVector, mip));
 
 #if defined(UNITY_USE_NATIVE_HDR) || defined(UNITY_DOTS_INSTANCING_ENABLED)
         irradiance += weightProbe1 * encodedIrradiance.rbg;
@@ -188,7 +188,7 @@ half3 CalculateIrradianceFromReflectionProbes(half3 reflectVector, float3 positi
     // Use any remaining weight to blend to environment reflection cube map
     if (totalWeight < 0.99f)
     {
-        half4 encodedIrradiance = half4(SAMPLE_TEXTURECUBE_LOD(_GlossyEnvironmentCubeMap, sampler_GlossyEnvironmentCubeMap, originalReflectVector, mip));
+        float4 encodedIrradiance = float4(SAMPLE_TEXTURECUBE_LOD(_GlossyEnvironmentCubeMap, sampler_GlossyEnvironmentCubeMap, originalReflectVector, mip));
 
 #if defined(UNITY_USE_NATIVE_HDR) || defined(UNITY_DOTS_INSTANCING_ENABLED)
         irradiance += (1.0f - totalWeight) * encodedIrradiance.rbg;
@@ -200,10 +200,10 @@ half3 CalculateIrradianceFromReflectionProbes(half3 reflectVector, float3 positi
     return irradiance;
 }
 
-half3 GlossyEnvironmentReflection(half3 reflectVector, float3 positionWS, half perceptualRoughness, half occlusion)
+float3 GlossyEnvironmentReflection(float3 reflectVector, float3 positionWS, float perceptualRoughness, float occlusion)
 {
 #if !defined(_ENVIRONMENTREFLECTIONS_OFF)
-    half3 irradiance;
+    float3 irradiance;
 
 #ifdef _REFLECTION_PROBE_BLENDING
     irradiance = CalculateIrradianceFromReflectionProbes(reflectVector, positionWS, perceptualRoughness);
@@ -212,8 +212,8 @@ half3 GlossyEnvironmentReflection(half3 reflectVector, float3 positionWS, half p
 #ifdef _REFLECTION_PROBE_BOX_PROJECTION
     reflectVector = BoxProjectedCubemapDirection(reflectVector, positionWS, unity_SpecCube0_ProbePosition, unity_SpecCube0_BoxMin, unity_SpecCube0_BoxMax);
 #endif // _REFLECTION_PROBE_BOX_PROJECTION
-    half mip = PerceptualRoughnessToMipmapLevel(perceptualRoughness);
-    half4 encodedIrradiance = half4(SAMPLE_TEXTURECUBE_LOD(unity_SpecCube0, samplerunity_SpecCube0, reflectVector, mip));
+    float mip = PerceptualRoughnessToMipmapLevel(perceptualRoughness);
+    float4 encodedIrradiance = float4(SAMPLE_TEXTURECUBE_LOD(unity_SpecCube0, samplerunity_SpecCube0, reflectVector, mip));
 
 #if defined(UNITY_USE_NATIVE_HDR)
     irradiance = encodedIrradiance.rgb;
