+++ PowerWater/PowerLib/WaveLib.hlsl
@@ -7,7 +7,7 @@ half3 GerstnerWave(half4 wave,half3 p,inout half3 tangent,inout half3 binormal){
     half k = 2 * PI / waveLength;
     half c = sqrt(9.8/k);
     half2 d = normalize(wave.xy);
-    half f = k * dot(d,p.xz) - c * _Time.y;
+    float f = k * dot(d,p.xz) - c * _Time.y;
     half a = steepness/k;
 
     tangent += half3(
